--- src/main/java/tech/dunny/model/Level.java
@@ -1,5 +1,0 @@
-package tech.dunny.model;
-
-public class Level {
-
-}
