+++ src/main/java/tech/teamfour/controllers/ScoreController.java
@@ -1,0 +1,4 @@
+package tech.teamfour.controllers;
+
+public class ScoreController {
+}
