+++ src/main/java/tech/teamfour/model/Level.java
@@ -1,0 +1,5 @@
+package tech.teamfour.model;
+
+public class Level {
+
+}
