+++ src/main/java/tech/teamfour/timer/Clock.java
@@ -1,4 +1,4 @@
-package tech.dunny.timer;
+package tech.teamfour.timer;
 
 import java.beans.PropertyChangeListener;
 import java.beans.PropertyChangeSupport;
