+++ src/main/resources/application.properties
@@ -6,5 +6,5 @@ spring.datasource.username=root
 spring.datasource.password=root
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.properties.hibernate.id.new_generator_mappings=true
-server.port=8082
+server.port=8081
 spring.jpa.open-in-view=false
