+++ src/test/java/tech/teamfour/cap_refactored/CapRefactoredApplicationTests.java
@@ -1,4 +1,4 @@
-package tech.dunny.cap_refactored;
+package tech.teamfour.cap_refactored;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
